split server go into seperate files
