--- server/server.go
@@ -1,269 +1,0 @@
-// Copyright (c) of parts are held by the various contributors (see the CLA)
-// Licensed under the MIT License. See LICENSE file in the project root for full license information.
-
-package server
-
-import (
-  "bytes"
-  "net/http"
-  "encoding/json"
-  "encoding/hex"
-  "github.com/pegnet/pegnet/opr"
-  "github.com/FactomProject/factom"
-)
-
-// RequestHandler as the base handler
-type RequestHandler struct{}
-
-// PostRequest struct to deserialise from request body
-type PostRequest struct {
-	Method string                 `json:"method"`
-	Params Parameters             `json:"params"`
-}
-
-// Parameters contains all possible json inputs
-type Parameters struct {
-  Address *string               `json:"address,omitempty"`
-  Height *int64                 `json:"height,omitempty"`
-  DigitalID string              `json:"digital-id,omitempty"`
-  Hash string                   `json:"hash,omitempty"`
-}
-
-// PostResponse to either contain a valid result or error
-type PostResponse struct {
-  Res   Result                  `json:"result"`
-	Err   Error                   `json:"error"`
-}
-
-// Result struct contains all potential json api responses
-type Result struct {
-  Balance int64                               `json:"balance,omitempty"`
-  Balances map[string]map[[32]byte]int64      `json:"balances,omitempty"`
-  ChainID string                              `json:"chainid,omitempty"`
-  LeaderHeight int64                          `json:"leaderheight,omitempty"`
-  OPRBlocks []*opr.OPRBlock                   `json:"oprblocks,omitempty"`
-  Winners []string                            `json:"winners,omitempty"`
-  Winner string                               `json:"winner,omitempty"`
-  OPRBlock *opr.OPRBlock                      `json:"oprblock,omitempty"`
-  OPRs []opr.OraclePriceRecord                `json:"oprs,omitempty"`
-  OPR opr.OraclePriceRecord                   `json:"opr,omitempty"`
-}
-
-// Error struct returns a code and it's associated message
-// 0: <Success>
-// 1: Method Not Found
-// 2: Parameter Not Found
-type Error struct {
-  Code    int                   `json:"code"`
-  Reason  string                `json:"reason"`
-}
-
-// Base handler of all requests
-func (h RequestHandler) ServeHTTP(writer http.ResponseWriter, req *http.Request) {
-  // log ip and request info
-  switch req.Method {
-	case "GET":
-		// serve dashboard
-	case "POST":
-		apiHandler(writer, req)
-	default:
-		methodNotAllowed(writer)
-	}
-}
-
-func apiHandler(w http.ResponseWriter, r *http.Request) {
-	var request PostRequest
-	json.NewDecoder(r.Body).Decode(&request)
-	// handle any decoder error and log
-
-	switch request.Method {
-	case "all-oprs":
-    response(w, Result{OPRBlocks: opr.OPRBlocks})
-    
-	case "balances":
-    response(w, Result{Balances: opr.Balances})
-    
-	case "balance":
-    getBalance(w, request.Params)
-
-  case "chainid":
-    response(w, Result{ChainID: opr.OPRChainID})
-    
-	case "leaderheight":
-		response(w, Result{LeaderHeight: getLeaderHeight()})
-
-  case "oprs-by-height":
-    getOPRsByHeight(w, request.Params)
-
-  case "oprs-by-id":
-    getOprsByDigitalID(w, request.Params)
-
-  case "opr-by-hash":
-    getOprByHash(w, request.Params)
-
-  case "opr-by-shorthash":
-    getOprByShortHash(w, request.Params)
-
-  case "winners":
-    winners := getWinners()
-    response(w, Result{Winners: winners[:]})
-
-  case "winner":
-    winner :=  getWinner()
-    response(w, Result{Winner: winner}) 
-
-  case "winning-opr":
-    winner :=  getWinner()
-    winningOPR := oprByShortHash(winner)
-    response(w, Result{OPR: winningOPR})
-
-	default:
-		errorResponse(w, Error{Code: 1, Reason: "Method Not Found"})
-	}
-}
-
-// getOprByHash handler to return the opr by full hash
-func getOprByHash(w http.ResponseWriter, params Parameters) {
-  if params.Hash != "" {
-    opr := oprByHash(params.Hash)
-    response(w, Result{OPR: opr})
-  } else {
-    invalidParameterError(w)
-  }
-}
-// getOprByShortHash handler to return the opr by the short 8 byte hash
-func getOprByShortHash(w http.ResponseWriter, params Parameters) {
-  if params.Hash != "" {
-    opr := oprByShortHash(params.Hash)
-    response(w, Result{OPR: opr})
-  } else {
-    invalidParameterError(w)
-  }
-}
-
-// getOprsByDigitalID handler to return all oprs based on Digital ID
-func getOprsByDigitalID(w http.ResponseWriter, params Parameters) {
-  if params.DigitalID != "" {
-    oprs := oprsByDigitalID(params.DigitalID)
-    response(w, Result{OPRs: oprs})
-  } else {
-    invalidParameterError(w)
-  }
-}
-
-// getOPRsByHeight handler will return all OPR's at any height except the current block.
-// Will only return local OPR's for the current chainhead.
-func getOPRsByHeight(w http.ResponseWriter, params Parameters) {
-  if params.Height !=nil {
-    oprblock := oprsByHeight(*params.Height)
-    response(w, Result{OPRBlock: oprblock})
-  } else {
-    invalidParameterError(w)
-  }
-}
-
-// getBalance handler will get the balance of a pegnet address
-func getBalance(w http.ResponseWriter, params Parameters) {
-  if params.Address != nil {
-    balance := opr.GetBalance(*params.Address)
-    res := Result{Balance: balance}
-    response(w, res)
-  } else {
-    invalidParameterError(w)
-  }
-}
-
-func methodNotAllowed(w http.ResponseWriter) {
-  http.Error(w, http.StatusText(405), http.StatusMethodNotAllowed)
-}
-
-// invalidParameterError returns when the method is valid but the parameter is not
-func invalidParameterError(w http.ResponseWriter) {
-  errorResponse(w,Error{Code: 2, Reason: "Parameter Not Found"})
-}
-
-// getWinners returns the current 10 winners entry shorthashes from the last recorded block
-func getWinners() [10]string {
-  height := getLeaderHeight()
-  currentOPRS := oprsByHeight(height)
-  opr := currentOPRS.OPRs[0]
-  return opr.WinPreviousOPR
-}
-
-// getWinner returns the highest graded entry shorthash from the last recorded block
-func getWinner() string {
-  return getWinners()[0]
-}
-
-
-// response is a wrapper around all responses to be served
-func response(w http.ResponseWriter, res Result){
-  json.NewEncoder(w).Encode(PostResponse{Res: res})
-}
-
-// errorResponse is a wrapper around all errors to be served
-func errorResponse(w http.ResponseWriter, err Error) {
-  json.NewEncoder(w).Encode(PostResponse{Err: err})
-}
-
-// getLeaderHeight helper function, cleaner than using the factom monitor
-func getLeaderHeight() int64 {
-  heights, err := factom.GetHeights()
-  if err != nil {
-    return 0
-  }
-  return heights.LeaderHeight
-}
-
-// OprsByHeight returns a single OPRBlock
-func oprsByHeight(dbht int64) *opr.OPRBlock {
-  for _, opr := range opr.OPRBlocks {
-    if opr.Dbht == dbht {
-      return opr
-    }
-  }
-  return nil
-}
-
-// OprsByDigitalID returns every OPR created by a given ID
-// Multiple ID's per miner or single daemon are possible.
-// This function searches searches and returns all.
-func oprsByDigitalID(did string) []opr.OraclePriceRecord {
-  var subset []opr.OraclePriceRecord
-  for _, block := range opr.OPRBlocks {
-    for _, opr := range block.OPRs{
-      for _, digitalID := range opr.FactomDigitalID{
-        if digitalID == did {
-          subset = append(subset, *opr)
-        }
-      }
-    }
-  }
-  return subset
-}
-
-// OprByHash returns the entire OPR based on it's hash 
-func oprByHash(hash string) opr.OraclePriceRecord {
-  for _, block := range opr.OPRBlocks {
-    for _, opr := range block.OPRs{
-      if hash == hex.EncodeToString(opr.OPRHash) {
-        return *opr
-      }
-    }    
-  }
-  return opr.OraclePriceRecord{}
-}
-
-// Failing tests. Need to grok how the short 8 byte winning oprhashes are done.
-func oprByShortHash(shorthash string) opr.OraclePriceRecord {
-  hashbytes, _  := hex.DecodeString(shorthash)
-    // hashbytes = reverseBytes(hashbytes)
-    for _, block := range opr.OPRBlocks {
-      for _, opr := range block.OPRs{
-        if bytes.Compare(hashbytes, opr.OPRHash[:8]) ==  0 {
-          return *opr
-        }
-      }    
-    }
-  return opr.OraclePriceRecord{}
-}
